requirements.txt
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
Django>=5.0.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.5.0
PyMySQL>=1.1.0
